everything work well

+++ src/main/java/by/borisevich/studentCatalog/controller/StudentServlet.java
@@ -12,6 +12,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.util.List;
 
@@ -28,40 +29,50 @@ public class StudentServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         logger.info("show Students get request");
-        int pageid = 1;
-        String spageid=req.getParameter("page");
-        if (spageid != null) {
-            pageid = Integer.parseInt(spageid);
+        HttpSession session = req.getSession();
+        if (session.getAttribute("user") != null) {
+            int pageid = 1;
+            String spageid = req.getParameter("page");
+            if (spageid != null) {
+                pageid = Integer.parseInt(spageid);
+            }
+            int recordsPerPage = 5;
+            List<Student> students = dao.getStudentsList((pageid - 1) * recordsPerPage + 1, recordsPerPage);
+            if (students != null) {
+                req.setAttribute("students", students);
+            }
+            int recordsCount = dao.getColOfRecords();
+            int noOfPages = (int) Math.ceil(recordsCount * 1.0 / recordsPerPage);
+            req.setAttribute("noOfPages", noOfPages);
+            req.setAttribute("currentPage", pageid);
+            req.getRequestDispatcher("home.jsp").forward(req, resp);
+        } else {
+            req.getRequestDispatcher("view/login.jsp").forward(req, resp);
         }
-        int recordsPerPage = 5;
-        List<Student>students = dao.getStudentsList((pageid - 1) * recordsPerPage + 1, recordsPerPage);
-        if (students != null) {
-            req.setAttribute("students", students);
-        }
-        int recordsCount = dao.getColOfRecords();
-        int noOfPages = (int) Math.ceil(recordsCount * 1.0 / recordsPerPage);
-        req.setAttribute("noOfPages", noOfPages);
-        req.setAttribute("currentPage", pageid);
-        req.getRequestDispatcher("home.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         logger.info("show Students post request");
-        String buttonValue = req.getParameter("studentButton");
-        String[] buttonArr = buttonValue.split("_");
-        String studentId = buttonArr[buttonArr.length-1];
-        if (buttonValue.contains("delStudent")) {
-            dao.deleteStudent(Integer.parseInt(studentId));
-            doGet(req,resp);
-        } else if (buttonValue.contains("editStudent")) {
-            List<String> groupsList = dao.getGroupNumList();
-            if (groupsList != null) {
-                req.setAttribute("groupsList", groupsList);
+        HttpSession session = req.getSession();
+        if (session.getAttribute("user") != null) {
+            String buttonValue = req.getParameter("studentButton");
+            String[] buttonArr = buttonValue.split("_");
+            String studentId = buttonArr[buttonArr.length - 1];
+            if (buttonValue.contains("delStudent")) {
+                dao.deleteStudent(Integer.parseInt(studentId));
+                doGet(req, resp);
+            } else if (buttonValue.contains("editStudent")) {
+                List<String> groupsList = dao.getGroupNumList();
+                if (groupsList != null) {
+                    req.setAttribute("groupsList", groupsList);
+                }
+                Student student = dao.getStudent(Integer.parseInt(studentId));
+                req = setReqParams(req, student);
+                req.getRequestDispatcher("view/addStudent.jsp").forward(req, resp);
             }
-            Student student = dao.getStudent(Integer.parseInt(studentId));
-            req = setReqParams(req, student);
-            req.getRequestDispatcher("view/addStudent.jsp").forward(req, resp);
+        } else {
+            req.getRequestDispatcher("view/login.jsp").forward(req, resp);
         }
     }
